Renamed main module to start and moved it into a directory
