+++ examples/paint_filter.rs
@@ -9,13 +9,13 @@ use winit::{
     event_loop::{ControlFlow, EventLoop},
 };
 
-mod main;
+mod start;
 
 fn main() {
     #[cfg(not(target_arch = "wasm32"))]
-    main::start(1000, 600, "Canvas::filter_image example", false);
+    start::start(1000, 600, "Canvas::filter_image example", false);
     #[cfg(target_arch = "wasm32")]
-    main::start();
+    start::start();
 }
 
 #[cfg(not(target_arch = "wasm32"))]
