+++ examples/svg.rs
@@ -6,13 +6,13 @@ use winit::{
     event_loop::{ControlFlow, EventLoop},
 };
 
-mod main;
+mod start;
 
 fn main() {
     #[cfg(not(target_arch = "wasm32"))]
-    main::start(1000, 600, "femtovg demo", true);
+    start::start(1000, 600, "femtovg demo", true);
     #[cfg(target_arch = "wasm32")]
-    main::start();
+    start::start();
 }
 
 #[cfg(not(target_arch = "wasm32"))]
