+++ examples/text.rs
@@ -12,13 +12,13 @@ struct Fonts {
     light: FontId,
 }
 
-mod main;
+mod start;
 
 fn main() {
     #[cfg(not(target_arch = "wasm32"))]
-    main::start(1000, 600, "Text demo", false);
+    start::start(1000, 600, "Text demo", false);
     #[cfg(target_arch = "wasm32")]
-    main::start();
+    start::start();
 }
 
 #[cfg(not(target_arch = "wasm32"))]
